When you render things with form_label, form_errors, form_widget and form_help you start to lose some of the special formatting that form_row() gives you. Sure, it's easy to re-add that div. But form_row also adds a special error class to that div when the field has a validation error.
For that reason, let's go back to using form_row().

A little bit later, we're going to learn how we can use form_row(), but completely customize how it looks for one specific form, or across your entire site! We'll do this by creating a 'form theme'. It's kind of the best of both worlds: you can render things in the lazy way, but still have the control you need.