We know what it looks like to create a new Article form: create the form, process the form request, and save the article to the database. But what does it look like to make an 'edit' form?
The answer is - delightfully - almost identical! In fact, let's copy all of our code from the new() action and go down to edit(), where the only thing we're doing so far is allowing Symfony to query for our article. Paste! Excellent.

Oh, but we need a few arguments: the Request and EntityManagerInterface $ em. This is now exactly the same code from the new form. So... how can we make this an edit form? You're going to love it! Pass $ article as the second argument to ->createForm().

We're done! Seriously! When you pass $ article, this object - which we just got from the database becomes the data attached to the form. This causes two things to happen. First, when Symfony renders the form, it calls the getter methods on that Article object and uses those values to fill in the values for the fields.

Heck, we can see this immediately! This is using the new template, but that's fine temporarily. Go to /article/1/edit. Dang - I don't have an article with id

Let's go find a real id. In your terminal, run:
php bin/console doctrine:query:sql 'SELECT * FROM article'
Perfect! Let's us id 26. Hello, completely pre-filled form!