Filling in the Form Type Extension
Ok! Let's remove the rest of the TODOs in here and then get to work! We can fill in whichever methods we need. In our case, we want to modify the view variables. That's easy for us: in buildView(), say $ view->vars['attr'], and then add a rows attribute equal to 10.

Done! Move over, refresh and... yea! I think it's bigger! Inspect it - yes: rows='10'. Every <textarea> on our entire site will now have this.