Validation Constraints
We do have some validation, for the password and agree to terms fields, because those constraints were added into the form class itself.
At this point, if you want to, you can remove these annotations from your User class. But, because we might use the User class on a form somewhere else - like an edit profile form - I'll keep them there.

One of the really nice things about using a form model class is that we can remove the constraints from the form and put them in the model class so that we have everything in one place. Above , add @Assert\NotBlank() and @Assert\Length(). Let's pass in the same options: message='' and copy that from the form class. Then copy the minMessage string, add min=5, minMessage= and paste.

Finally, above agreeTerms, go copy the message from the form, and add the same @Assert\IsTrue() with message= that message.